% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Actions performed when \code{library(YehLabHTS)} is called.}
\usage{
.onLoad(libname, pkgname)
}
\description{
This file assigns our metadata and raw plate data objects to the global environment upon loading the package.
}
