% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBlissMatrix.R
\name{generateBlissMatrix}
\alias{generateBlissMatrix}
\title{A function to generate bliss matrices for anchor / library drug combinations}
\usage{
generateBlissMatrix(drug.data)
}
\arguments{
\item{drug.data}{The input anchor / drug data.}
}
\description{
This function, given an input dataframe containing info on library dose, anchor dose, and inhibition, generates a list of bliss matrices for each combination. The matrix is intended to be later used as input to a heatmap function.
}
\examples{
generateBlissMatrix(drug.data = my_data)
}
