% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatMetadata.R
\name{reformatMetadata}
\alias{reformatMetadata}
\title{Reformat metadata}
\usage{
reformatMetadata()
}
\description{
This function reformats the metadata and library sheets from out HTS experiments into a long format.
}
\examples{
reformatMetadata()
}
