% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHTSData.R
\name{parseHTSData}
\alias{parseHTSData}
\title{Parse raw HTS plate data}
\usage{
parseHTSData(seed.use = 629)
}
\arguments{
\item{seed.use}{The random seed used to control stochasticity. Defaults to 629.}
}
\description{
This functions imports HTS data in \code{.xlsx} format, parses it, and prepares it for synergy analysis.
}
\examples{
parseHTSData()
}
